# server/docs/swagger.yaml
basePath: /api
definitions:
...
  exercise.ExerciseHistorical1RMResponse:
    properties:
      computed_best_e1rm:
        type: number
      computed_best_workout_id:
        type: integer
      historical_1rm:
        type: number
      historical_1rm_source_workout_id:
        type: integer
      historical_1rm_updated_at:
        type: string
    type: object
...
  exercise.UpdateExerciseHistorical1RMRequest:
    properties:
      historical_1rm:
        maximum: 999999.99
        minimum: 0
        type: number
      mode:
        enum:
        - manual
        - recompute
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      message:
        example: Error message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A fitness tracking application API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FitTrack API
  version: "1.0"
paths:
...
  /exercises/{id}/historical-1rm:
    get:
      consumes:
      - application/json
      description: Get stored historical 1RM metadata for an exercise, plus a computed
        best e1RM suggestion from working sets.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.ExerciseHistorical1RMResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found - Exercise not found or doesn't belong to user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Get exercise historical 1RM
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      description: Set a manual historical 1RM (clears source workout) or recompute
        from best working-set e1RM across history.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      - description: Historical 1RM update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/exercise.UpdateExerciseHistorical1RMRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Exercise historical 1RM updated successfully
        "400":
          description: Bad Request - Invalid exercise ID or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized - Invalid token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found - Exercise not found or doesn't belong to user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Update exercise historical 1RM
      tags:
      - exercises
...
schemes:
- http
securityDefinitions:
  StackAuth:
    in: header
    name: x-stack-access-token
    type: apiKey
swagger: "2.0"